requirements.txt
wikipedia
pyjokes
pillow
SpeechRecognition==3.8.1
PyAudio
pyttsx3

